\name{evidence}
\alias{evidence}
\title{Short evidence representation}
\description{The function converts long RegulonDB strings about the evidence into 
short and clear representation with letters |S| and |W|. }
\usage{evidence(evidence.string)}
\arguments{
  \item{evidence.string}{RegulonDB string, containing the evidence for a TFBS or a promoter}
}



\author{Evgenia Temlyakova}





