\name{tfbs.pos}
\alias{tfbs.pos}
\title{TFBS positions on the E.coli chromosome}
\description{The function locates all binding sites for specified TF on E.coli chromosome for one or both DNA strands. It uses information taken from RegulonDB version 8.2.}
\usage{tfbs.pos(tfname, strand = "both")}
\arguments{
  \item{tfname}{a name of TF}
  \item{strand}{
}
}



\author{Evgenia Temlyakova}





