\name{TFBS.QSAM-package}
\alias{TFBS.QSAM-package}
\alias{TFBS.QSAM}
\docType{package}
\title{The analysis of TFBS by QSAM}
\description{The package provides various techniques to analyse physical and chemical properties of DNA sequencies. The most attention is given to transription factor binding sites.}
\details{
\tabular{ll}{Package: \tab TFBS.QSAM\cr
Type: \tab Package\cr
Title: \tab The analysis of TFBS by QSAM\cr
Version: \tab 1.0.1\cr
Date: \tab 2013-12-04\cr
Author: \tab Evgenia Temlyakova\cr
Maintainer: \tab Evgenia Temlyakova <evgenia.teml@gmail.com>\cr
License: \tab GPL (>= 2)\cr}
}
\author{Evgenia Temlyakova}

\keyword{ package }


